[♻️refactor]: 컬러값과 그 외 디자인 시스템 분리
- 컬러 외에는 ThemeProvider의 영향을 받지 않음
- 따라서 컬러 값을 제외한 값들은 DESIGN_SYSTEM이라는 객체에 저장함
